Implemented writing time dimension from runlog
I'm also starting to refactor things out of the `interface` module;
eventually I'd like to get rid of that and move all the components
to their respective modules.

Additionally, I found an inconsistency in how lifetimes are defined
for one of the `netcdf::GroupMut` methods, `add_variable_with_type`.
This was causing a problem for me where I couldn't pass a mutable
group reference into a function and call this method without extra
lifetime bounds, which then meant I couldn't use the reference in
successive iterations of a for loop. For now, I've forked the
netcdf repo and pointed my Cargo.toml to it, but I'd like to submit
an issue on this at some point.