Started work on public netCDF writer
Mainly implemented PEST parsing for template strings to use in the config.
My current plan is that the config can either specify 1:1 variable mapping,
a group where the private names are spelled out but the public names and
attributes can be done programmatically, or a wildcard group (maybe) where
we specify a pattern for private variables.