Added effective vertical path test program and data
I've created a Fortran program that calls to `compute_vertical_paths.f`
in `$GGGPATH/src/gfit` to generate a set of effective vertical paths
to use in testing the Rust version of the same subroutine. The path
data is in `vpath.dat` so it only needs recreated if `compute_vertical_paths`
changes.