Checkpointing progress towards putting variables in groups correctly.
I've implemented a `GroupSelector` that reads in the prefixes file, and
also changed that file to provide columns for the suffix and group name.
The standard group enum now just differentiates between the main group
and any experimental group, with the experimental group storing its
group name and suffix. Variables-to-be also now hold the group they are
to go in, making it easier to write to multiple groups from one
provider.

However, I need to consider exactly how variables will be divided into
groups over the whole post processing chain. My testing with an existing
Lamont run directory is having some issues because not all variables are
prefixed with "m" or "v" for InSb and Si data, respectively. I think it
would be simpler to use the approach that I think
`collate_tccon_results` does, which is to prefix everything from the
secondary detectors with the appropriate prefix. That would mean I need
to rerun post processing on a Lamont run directory to test this out.

We may also rethink exactly how variables get split up; Debra suggested
simply defining a list of variables that we consider standard and
putting everything else in a single "experimental" group. That would
take the InSb/Si terminology out of the files, which makes sense -
that's a very internal detail. Because of that discussion, I would like
to change the prefix definition file to a TOML file so that it can
easily have different sections, e.g.:

- define wavenumber groups (as we have done),
- point to a window file that defines the standard windows, or
- directly define the standard windows in that file.

Something to consider is how defining a list of standard windows would
interact with the need to prefix variable names to ensure
`average_results` doesn't merge then together. It might be that
`collate_tccon_results` requires the wavenumber definitions, and the netCDF
writer could use any definitions.