Reworked TCCON post processing so that O2 DMFs are in the files from the start
I had previous set things up so that `apply_tccon_airmass_correction` would be where
the O2 DMFs used in computing the Xgas values would be gotten. After all, that is where
the value is needed. However, since we want to track the O2 DMF in the post processing
files, this meant that we would add an auxiliary column after `apply_tccon_airmass_correction`.
That makes serializing and deserializing the post processing data awkward, since we have to
allow for that value to be absent, and serializing a `None` with Fortran formatting is just
weird.

Instead, `collate_tccon_results` now inserts the O2 DMF value into the `.Xsw` files. Since it
can just use a fixed value, this shouldn't be an imposition for users that don't care about
Xgas values. This allows us to keep the O2 DMF as a value in the `AuxData` struct, without it
being optional. With this, I also reworked how `AuxData` structs are constructed to follow the
builder pattern - this way there aren't random fill values in the structure until we remember
to replace them.