Removed default aux vars and added EM27 included config
After working with the public writer for a bit, I realized that it is more confusing to have
the default auxiliary variables be baked into the code, rather than in the configuration.
That was edging towards the same sort of duplication of concerns that made the previous
public writer hard to maintain. Now, I've moved those into a "common" configuration,
which gets included in the standard TCCON and EM27 configurations.

I've also added an EM27 standard configuration, based on testing with Toronto EM27/SUN
data.